tests/install_startup.pm: disable UEFI boot for legacy
For legacy, it isn't enough to choose proper boot medium for the
installer, because on next boot platform would attempt booting UEFI
first, which would land in UEFI installer. To make it work, we could
either remove the installer medium (on PiKVM) or disable UEFI booting.
Of course, it must be re-enabled for UEFI installation.

The test now also ensures that USB devices are attempted after SATA
in order to allow booting from disk without disconnecting the drive.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>