tests/install_startup.pm: generate kickstart config from common template
Kickstart files are now created from a common template, with potential
workarounds applied as needed. They are served from per-worker location
to avoid conflicts with concurrent jobs.

This change also fixes bug in workaround for HP T630 that wasn't caught.
'sed' was missing the switch to do in-place modification. As a result,
patched version was only printed on stdout and not saved in any file.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>