generalhw/*/ks.cfg: remove files and references to them in flash script
Supermicro is left mostly unchanged, except for comments, but its ks.cfg
was removed. It is expected to fail, but the platform is unable to launch
AEM at a hardware level.

This change turns 'flash' scripts for hpt630v1, vp4670 and optiplex into
no-ops. It also removes 'ipxe' script for hpt630v1, as it was never used.

'ipxe' scripts for other platforms will require further changes to point
to crafted 'ks.cfg' files, this will be done in later commits. The same
applies to 'QUBES_OS_OPENQA_URL' worker option. Both are used only for
supermicro or legacy boot on optiplex. As an alternative, a local boot
may be attempted, similarly to how it is done on hpt630v1.

'QUBES_OS_KS_URL' option can be removed from 'workers.ini' for supported
platforms, but the code for handling it was left untouched. This path
would be taken by msi, but it wasn't tested in a long time and may not
work currently.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>